Implement deft method

'deft' method is a meta-analytical approach to pool conclusion from multiple studies. More details please see references.

deft_do(prepare, group_level, method = "FE")

Arguments

prepare
a result data.frame from deft_prepare function or a data.frame contains at least 'trial', 'subgroup', 'yi' and 'sei' these four columns.

group_level
level of subgroup, should be a character vector with length 2 and the reference should put in the first. For example, if you have 'Male' and 'Female' groups and want compare 'Female' with 'Male', then should set c('Male', 'Female').

method
character string specifying whether a fixed- or a random/mixed-effects model should be fitted. A fixed-effects model (with or without moderators) is fitted when using method="FE". Random/mixed-effects models are fitted by setting method equal to one of the following: "DL", "HE", "SJ", "ML", "REML", "EB", "HS", or "GENQ". Default is "REML". See ‘Details’.

Value

a list which class is 'deft'.

Details

About model fit, please see metafor::rma().

References

Fisher, David J., et al. "Meta-analytical methods to identify who benefits most from treatments: daft, deluded, or deft approach?." bmj 356 (2017): j573.

Wang, Shixiang, et al. "The predictive power of tumor mutational burden in lung cancer immunotherapy response is influenced by patients' sex." International journal of cancer (2019).

Examples

data('wang2019')
deft_do(wang2019, group_level = c("Male", "Female"))

#> $all
#> $all$data
#>                  entry         trial subgroup   hr ci.lb ci.ub  ni   conf_q
#> 1      Rizvi 2015-Male    Rizvi 2015     Male 0.30  0.09  1.00  16 1.959964
#> 2    Rizvi 2015-Female    Rizvi 2015   Female 0.11  0.02  0.56  18 1.959964
#> 3      Rizvi 2018-Male    Rizvi 2018     Male 1.25  0.82  1.90 118 1.959964
#> 4    Rizvi 2018-Female    Rizvi 2018   Female 0.63  0.42  0.95 122 1.959964
#> 5   Hellmann 2018-Male Hellmann 2018     Male 0.90  0.41  1.99  37 1.959964
#> 6 Hellmann 2018-Female Hellmann 2018   Female 0.28  0.12  0.67  38 1.959964
#>           yi       sei
#> 1 -1.2039728 0.6142831
#> 2 -2.2072749 0.8500678
#> 3  0.2231436 0.2143674
#> 4 -0.4620355 0.2082200
#> 5 -0.1053605 0.4030005
#> 6 -1.2729657 0.4387290
#> 
#> $all$model
#> 
#> Fixed-Effects Model (k = 6)
#> 
#> Test for Heterogeneity:
#> Q(df = 5) = 18.8865, p-val = 0.0020
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval    ci.lb    ci.ub 
#>  -0.3207  0.1289  -2.4883  0.0128  -0.5733  -0.0681  * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> 
#> 
#> $subgroup
#> $subgroup$data
#>           trial        hr     ci.lb     ci.ub  ni   conf_q        yi       sei
#> 1 Hellmann 2018 0.3111111 0.0967900 1.0000013  75 1.959964 -1.167605 0.5957285
#> 2    Rizvi 2015 0.3666667 0.0469397 2.8641945  34 1.959964 -1.003302 1.0487893
#> 3    Rizvi 2018 0.5040000 0.2805772 0.9053338 240 1.959964 -0.685179 0.2988460
#> 
#> $subgroup$model
#> 
#> Fixed-Effects Model (k = 3)
#> 
#> Test for Heterogeneity:
#> Q(df = 2) = 0.5657, p-val = 0.7536
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval    ci.lb    ci.ub 
#>  -0.7956  0.2589  -3.0737  0.0021  -1.3030  -0.2883  ** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> 
#> 
#> attr(,"class")
#> [1] "deft"